trailing comma; resolved issue with loading more after altering slider
I introduced a variable for backing up the value of the integer minimum rating. This way, if you try to modify the slider, then press load more, you would have movies loaded from the same criteria you generated previously (the slider resets to the previous value).  The reason for this is that creating a new criteria implies you want to generate a new list; you wouldn't want to generate a page of min 9.0 movies and, below it, one of min 5.0.